Type deserialization definition
