+++ src/compiler/deserialization/mod.ts
@@ -2,3 +2,4 @@ export * from './default.ucd-defs.js';
 export * from './ucd-def.js';
 export * from './ucd-function.js';
 export * from './ucd-lib.js';
+export * from './ucd-type-def.js';
